Data retrieval from South Florida Water Management District online database (DBHYDRO)

Data retrieval from South Florida Water Management District online database (DBHYDRO)

DBHYDRO_daily(
  SDATE,
  EDATE,
  DBK,
  dataonly = TRUE,
  period = "uspec",
  v_target_code = "file_csv",
  vert_datum = 1,
  ...
)

Arguments

SDATE

input start date as.Date (YYYY-MM-DD)

EDATE

input end date as.Date (YYYY-MM-DD)

DBK

SFWMD daily data site and data type identifier

dataonly

DEFAULT is TRUE to return a data.frame of data, if set to FALSE it will return a nested list of two data.frame labeled METADATA and REPORT.

period

DEFAULT set to "uspec" for URL query

v_target_code

DEFAULT set to "file_csv" to return a CSV file

vert_datum

DEFAULT set to 1 for NGVD29 vertical datum; 2 = NAVD88 vertical datum

...

to allow for more functionality and flexibility in building URL queries

Value

This function returns daily hydrometerological (discharge, stage(WL) and meterological parameters) dataset from the SFWMD monitoring network (https://apps.sfwmd.gov/WAB/EnvironmentalMonitoring/index.html). This function assumes some familiarity with the District monitoring network and data management.

Examples

if (FALSE) {
# Daily Discharge Data
sdate=as.Date("2001-05-01");
edate=as.Date("2002-05-01");
dat=DBHYDRO_daily(sdate,edate,"FE771")
}