Water Quality Data retrieval from South Florida Water Management District online database (DBHYDRO)

using the new data management platform DBHydro Insights, this function using the Districts API access to retrieve data from the database

insight_fetch_wq(
  startDate,
  endDate,
  station_id,
  test_number,
  methods = NULL,
  projects = NULL,
  matrices = NULL,
  paramGroups = NULL,
  sampleTypes = NULL,
  reportType = "timeseries",
  format = "csv",
  ...
)

Arguments

startDate

Required as a Date as.Date or other format (i.e. as.POSIXct)

endDate

Required as a Date as.Date or other format (i.e. as.POSIXct)

station_id

SFWMD Water Quality Monitoring location https://insights.sfwmd.gov/#/mappage/site

test_number

The DBHydro 1 to 4 digit test number, see details

methods

A 1 to 3 letter code for the method of sample collection, see details

projects

DBHydro water quality project codes

matrices

A 1 to 3 letter code for the sampling material

paramGroups

Letter code for the parameter group, see details

sampleTypes

A letter code defining sample and other type of samples (i.e. QC/QC)

reportType

default is timeseries

format

default is csv

...

not used, to pass argument along to other functions, in the future.

Details

For test_number, methods,projects,matrices,sampleTypes and others see the insight_ref() function for the corresponding tables.

For paramGroups

O: organics

N: nutrients

M: metals

P: physical parameters

B: biological

F: field

MI: major ions

MIS: miscellaneous

Examples

if (FALSE) {
sdate <- as.Date("2001-05-01");
edate <- as.Date("2002-05-01");
dat <- insight_fetch_wq(sdate,edate,c("S12A","S12B"),"25")

# retrieve metadata
attr(dat,"metadata")
}